remove old ServiceSingle and rename ServicePage to ServiceSingle
